refactor: 전역 스타일 이외의 스타일을 컴포넌트 내부에 넣음
