--- styles/NotificationButton/Display.ts
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-const Display = styled.div`
-  background-color: #ff924a;
-  border-radius: 50%;
-  width: 8px;
-  height: 8px;
-  position: absolute;
-  transform: translate(0%, -50%);
-  left: 19px;
-  top: 13px;
-  z-index: 1;
-`;
-
-export default Display;
