--- styles/NotificationButton/Frame.ts
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-const Frame = styled.div`
-  background-color: rgba(255, 255, 255, 0.45);
-  border-radius: 10px;
-  width: 36px;
-  height: 36px;
-  position: absolute;
-  display: inline-block;
-  right: 0;
-  transform: translateY(-50%);
-  top: 50%;
-`;
-
-export default Frame;
