--- styles/NotificationButton/Icon.ts
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-const Icon = styled.svg`
-  width: 17px;
-  height: 17px;
-  fill: white;
-  position: absolute;
-  transform: translate(-50%, -50%);
-  left: 50%;
-  top: 50%;
-`;
-
-export default Icon;
