--- styles/NotificationButton/index.ts
@@ -1,11 +1,0 @@
-import Display from "./Display";
-import Frame from "./Frame";
-import Icon from "./Icon";
-
-const NotificationButton = {
-  Frame: Frame,
-  Icon: Icon,
-  Display: Display,
-};
-
-export default NotificationButton;
