--- styles/PageTitle/Title.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-const Title = styled.a`
-  font-size: 30px;
-  font-weight: bold;
-  color: white;
-`;
-
-export default Title;
