--- styles/PageTitle/index.ts
@@ -1,7 +1,0 @@
-import Title from "./Title";
-
-const PageTitle = {
-  Title: Title,
-};
-
-export default PageTitle;
