--- styles/PlantingGlass/Card.ts
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const Card = styled.div`
-  background-color: white;
-  padding: 3% 4%;
-  box-sizing: border-box;
-  box-shadow: 0px 0px 3px rgba(0, 0, 0, 0.3);
-  border-radius: 10px;
-  height: 110px;
-`;
-
-export default Card;
