--- styles/PlantingGlass/Grid.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-const Grid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(24, 3%);
-  grid-gap: 1.215%;
-`;
-
-export default Grid;
