--- styles/PlantingGlass/WeekItem.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-const WeekItem = styled.div`
-  display: flex;
-`;
-
-export default WeekItem;
