--- styles/PlantingGlass/index.ts
@@ -1,15 +1,0 @@
-import Card from "./Card";
-import DayItem from "./DayItem";
-import Grid from "./Grid";
-import WeekGrid from "./WeekGrid";
-import WeekItem from "./WeekItem";
-
-const PlantingGlass = {
-  Card: Card,
-  Grid: Grid,
-  WeekGrid: WeekGrid,
-  WeekItem: WeekItem,
-  DayItem: DayItem,
-};
-
-export default PlantingGlass;
