--- styles/RepresentativeLayout/ContentsSection.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const ContentsSection = styled.div`
-  margin-bottom: 90px;
-  overflow: visible;
-`;
-
-export default ContentsSection;
