--- styles/RepresentativeLayout/HeaderSection.ts
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const HeaderSection = styled.div`
-  height: 55px;
-  line-height: 55px;
-  position: sticky;
-  top: 10px;
-  margin-top: 10px;
-  z-index: 3;
-`;
-
-export default HeaderSection;
