--- styles/RepresentativeLayout/index.ts
@@ -1,11 +1,0 @@
-import ContentsSection from "./ContentsSection";
-import HeaderSection from "./HeaderSection";
-import HeaderBackground from "./HeaderBackground";
-
-const RepresentativeLayout = {
-  HeaderSection: HeaderSection,
-  HeaderBackground: HeaderBackground,
-  ContentsSection: ContentsSection,
-};
-
-export default RepresentativeLayout;
