--- styles/TodayLayout/PlantingGlassSection.ts
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-const PlantingGlassSection = styled.div`
-  max-width: 350px;
-  height: 120px;
-  margin: auto;
-  margin-top: 10px;
-  overflow: visible;
-  position: sticky;
-  top: 75px;
-  z-index: 2;
-`;
-
-export default PlantingGlassSection;
