--- styles/TodayLayout/TodoListSection.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-const TodoListSection = styled.div`
-  max-width: 350px;
-  margin: auto;
-  margin-top: 10px;
-  overflow-x: visible;
-  z-index: -1;
-`;
-
-export default TodoListSection;
