--- styles/TodayLayout/index.ts
@@ -1,9 +1,0 @@
-import PlantingGlassSection from "./PlantingGlassSection";
-import TodoListSection from "./TodoListSection";
-
-const RepresentativeLayout = {
-  TodoListSection: TodoListSection,
-  PlantingGlassSection: PlantingGlassSection,
-};
-
-export default RepresentativeLayout;
