--- styles/TodayTodoItem/ButtonSection.ts
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const ButtonSection = styled.div`
-  width: 120px;
-  position: absolute;
-  right: 0;
-  transform: translateY(-50%);
-  top: 50%;
-  z-index: -10;
-`;
-
-export default ButtonSection;
