--- styles/TodayTodoItem/CourseTitle.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-const CourseTitle = styled.a`
-  font-family: "NanumBarunGothic", "serif";
-  font-size: 13px;
-  /* font-weight: bold; */
-  color: #afafaf;
-  margin-left: 28px;
-`;
-
-export default CourseTitle;
