--- styles/TodayTodoItem/EventButton.ts
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-const EventButton = styled.button`
-  width: 50px;
-  height: 50px;
-  margin: 5px;
-  border-radius: 50%;
-  border: 0px;
-  box-shadow: 0px 0px 3px rgba(0, 0, 0, 0.3);
-  background-color: white;
-  position: relative;
-`;
-
-export default EventButton;
