--- styles/TodayTodoItem/TodoItemSection.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-const TodoItemSection = styled.div`
-  width: 100%;
-  height: 55px;
-  margin-bottom: 10px;
-  position: relative;
-  z-index: 0;
-`;
-
-export default TodoItemSection;
