--- styles/TodayTodoItem/TodoTitle.ts
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-const TodoTitle = styled.a`
-  font-family: "NanumBarunGothic", "serif";
-  font-size: 14px;
-  color: #747474;
-  font-weight: bold;
-  margin-left: 10px;
-
-  position: absolute;
-  transform: translateY(-50%);
-  top: 56%;
-`;
-
-export default TodoTitle;
