--- styles/TodayTodoItem/TodoTitleSection.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const TodoTitleSection = styled.div`
-  height: 18px;
-  position: relative;
-`;
-
-export default TodoTitleSection;
