--- styles/TodayTodoItem/index.ts
@@ -1,23 +1,0 @@
-import Icon from "./Icon";
-import ButtonSection from "./ButtonSection";
-import CourseColor from "./CourseColor";
-import CourseTitle from "./CourseTitle";
-import EventButton from "./EventButton";
-import TodoItemCard from "./TodoItemCard";
-import TodoItemSection from "./TodoItemSection";
-import TodoTitle from "./TodoTitle";
-import TodoTitleSection from "./TodoTitleSection";
-
-const TodayTodoItem = {
-  TodoTitleSection: TodoTitleSection,
-  CourseColor: CourseColor,
-  TodoItemCard: TodoItemCard,
-  TodoTitle: TodoTitle,
-  CourseTitle: CourseTitle,
-  TodoItemSection: TodoItemSection,
-  ButtonSection: ButtonSection,
-  EventButton: EventButton,
-  Icon: Icon,
-};
-
-export default TodayTodoItem;
